% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_colnames.R
\name{align_dict_to_schema}
\alias{align_dict_to_schema}
\title{Harmonise dict names to schema headers}
\usage{
align_dict_to_schema(dict_names, xml_path)
}
\arguments{
\item{dict_names}{character vector from your Excel dict}

\item{xml_path}{path to one representative schema XML}
}
\value{
named character vector: dict_name -> matched schema header
}
\description{
Harmonise dict names to schema headers
}
