% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct_assign_system.R
\name{assign_system}
\alias{assign_system}
\title{SNOMED System Classifier}
\usage{
assign_system(term)
}
\arguments{
\item{term}{A character string (the concept label or synonym)}
}
\value{
A character string containing the matching system(s) separated by "|",
or "Vague" if no system matches
}
\description{
Assign one or more body‑system labels to a term
}
