% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurum_make_denom_db.R
\name{aurum_make_denom_db}
\alias{aurum_make_denom_db}
\title{make a demographic linkage database for control construction}
\usage{
aurum_make_denom_db(dbn, denomdir, linkdir, db = F)
}
\description{
make a demographic linkage database for control construction
}
