% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_with_data_dictionary.R
\name{construct_col_types}
\alias{construct_col_types}
\title{Construct a col_type parameter from the data-dictionary to pass to the readr::read_tsv function}
\usage{
construct_col_types(data_dict)
}
\description{
Construct a col_type parameter from the data-dictionary to pass to the readr::read_tsv function
}
