% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmplud_make_db_py.R
\name{dmplusd_make_db_py}
\alias{dmplusd_make_db_py}
\title{make a duckdb}
\usage{
dmplusd_make_db_py(
  filePath,
  dbPath = dirname(filePath),
  dbName = NULL,
  xml_config = system.file("xml", "dmd_structure.xml", package = "rtrhd"),
  db = F,
  ow = F
)
}
\description{
make a duckdb
}
