% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmplud_make_db_r.R
\name{dmplusd_make_db_r}
\alias{dmplusd_make_db_r}
\title{make a duckdb}
\usage{
dmplusd_make_db_r(
  filePath,
  dbPath = dirname(filePath),
  dbName = NULL,
  wks = 3,
  db = F,
  ow = F
)
}
\description{
make a duckdb
}
