% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sankey_data.R
\name{gen_sankey_data}
\alias{gen_sankey_data}
\title{Generate Sankey Network Data}
\usage{
gen_sankey_data(
  dbf,
  condition = "SLE",
  status = "child",
  tabstub = paste0("temp_", condition),
  tabname = paste0(condition, "_sankey"),
  db = F
)
}
\description{
This is necessary because of the limitations on SQL Statements to duckdb
}
