% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_periods.R
\name{generate_periods}
\alias{generate_periods}
\title{generate set of regular periods}
\usage{
generate_periods(months_before, months_after, period_length, aslist = T)
}
\description{
generate set of regular periods
}
