% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snomed_ancestors.R
\name{get_snomed_ancestors}
\alias{get_snomed_ancestors}
\title{Get SNOMED Ancestors from a Database}
\usage{
get_snomed_ancestors(
  concept_id,
  db_path,
  concept_table = "terminology_concept",
  relationship_table = "terminology_relationship",
  description_table = "terminology_description",
  summarise = FALSE
)
}
\arguments{
\item{concept_id}{SNOMED concept ID}

\item{db_path}{Path to DuckDB database}

\item{concept_table}{Name of concept table}

\item{relationship_table}{Name of relationship table}

\item{description_table}{Name of description table}

\item{summarise}{Logical; if TRUE return semantic summary}
}
\value{
A data frame of ancestors or a summarised version by semantic type
}
\description{
Retrieves SNOMED ancestors and optionally summarises terms by semantic type.
}
