% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_make_denom_db.R
\name{gold_make_denom_db}
\alias{gold_make_denom_db}
\title{make a demographic linkage database for control construction}
\usage{
gold_make_denom_db(dbn, denomdir, linkdir, db = F)
}
\description{
make a demographic linkage database for control construction
}
