% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_make_db.R
\name{icd_make_db}
\alias{icd_make_db}
\title{make a snomed database}
\usage{
icd_make_db(
  dbPath,
  txtPath,
  dbName = NULL,
  db = F,
  ow = T,
  preterm = NULL,
  posterm = NULL,
  prerefs = NULL,
  posrefs = NULL
)
}
\description{
Trying to use some of my existing convenience functions
}
