% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_aurum_lookups.R
\name{load_aurum_lookups}
\alias{load_aurum_lookups}
\title{load the aurum lookup tables}
\usage{
load_aurum_lookups(pddir, dbf, ow = F, db = F, silent = T, prefix = "ref")
}
\description{
Pass a table of covariate codes and generate covariates table
}
