% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_aurum_observations.R
\name{load_aurum_observations}
\alias{load_aurum_observations}
\title{get the drugissue records and convert some fields to useful field types}
\usage{
load_aurum_observations(
  pddir,
  dbf,
  ow = F,
  db = F,
  tab_name = "observations",
  add = F,
  selvars = c("patid", "consid", "pracid", "obsid", "obsdate", "enterdate", "staffid",
    "parentobsid", "obstypeid", "medcodeid", "obsdate", "value", "numunitid",
    "numrangelow", "numrangehigh", "probobsid")
)
}
\arguments{
\item{bdir}{the name of the batch subdirectory}

\item{odir}{the name of the output directory}

\item{olist}{the list of observation codes}

\item{bpp}{BiocParallal Multicore Parameters

Pass a table of covariate codes and generate covariates table}
}
\description{
get the drugissue records and convert some fields to useful field types
}
