% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_aurum_patients.R
\name{load_aurum_patients}
\alias{load_aurum_patients}
\title{load patients}
\usage{
load_aurum_patients(
  pdir,
  dbf,
  ow = F,
  db = F,
  tab_name = "patients",
  add = F,
  selvars1 = c("patid", "pracid", "gender", "yob", "regstartdate", "regenddate",
    "emis_ddate", "cprd_ddate")
)
}
\arguments{
\item{batch}{the name of the batch subdirectory}

\item{datadir}{the name of the data directory}
}
\description{
get the patient records and convert some fields to useful field types
}
