% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_aurum_practices.R
\name{load_aurum_practices}
\alias{load_aurum_practices}
\title{get observations batch}
\usage{
load_aurum_practices(
  pddir,
  dbf,
  ow = F,
  db = F,
  tab_name = "practices",
  add = F,
  selvars = c("pracid", "lcd", "uts", "region")
)
}
\arguments{
\item{bdir}{the name of the batch subdirectory}

\item{odir}{the name of the output directory}

\item{olist}{the list of observation codes}

\item{bpp}{BiocParallal Multicore Parameters

Pass a table of covariate codes and generate covariates table}
}
\description{
get the drugissue records and convert some fields to useful field types
}
