% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_with_data_dictionary.R
\name{load_file_with_data_dictionary}
\alias{load_file_with_data_dictionary}
\title{Find all files that match the sheet name pattern and read and load them to the database}
\usage{
load_file_with_data_dictionary(
  fname,
  ddict,
  ddir,
  dbf,
  proto,
  dset,
  reppref = F,
  ow = F,
  db = F
)
}
\description{
Find all files that match the sheet name pattern and read and load them to the database
}
