% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gold_lookups.R
\name{load_gold_lookups}
\alias{load_gold_lookups}
\title{load cprd files}
\usage{
load_gold_lookups(
  pddir,
  dbf,
  ow = T,
  db = F,
  inctxt = "TXTFILES",
  dropcodes = T,
  prefix = "gold"
)
}
\arguments{
\item{bdir}{the name of the batch subdirectory}

\item{odir}{the name of the output directory}

\item{olist}{the list of observation codes}

\item{bpp}{BiocParallal Multicore Parameters

Pass a table of covariate codes and generate covariates table}
}
\description{
get the drugissue records and convert some fields to useful field types
}
