% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct_make_db.R
\name{load_sct_file}
\alias{load_sct_file}
\title{Load a single SNOMED CT file into DuckDB}
\usage{
load_sct_file(tab, con, filelist, ow = FALSE)
}
\arguments{
\item{tab}{Table name.}

\item{con}{An open DBI connection to the target DuckDB database.}

\item{filelist}{Named list of filenames.}

\item{ow}{Overwrite if table exists? (default FALSE)}
}
\value{
NULL (invisible), or result from load_table
}
\description{
Loads one file into a DuckDB table, skipping or overwriting as specified.
}
