% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_table.R
\name{load_table}
\alias{load_table}
\title{Load a table}
\usage{
load_table(
  filename = NULL,
  dataset = NULL,
  dbf = NULL,
  con = NULL,
  ow = F,
  db = F,
  append = F,
  tab_name = gsub("(^[0-9]*_|[.].*)", "", basename(filename)),
  selvars = NULL,
  delim = "\\t",
  quote = ""
)
}
\description{
Load a table
}
