% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_with_data_dictionary.R
\name{load_with_data_dictionary}
\alias{load_with_data_dictionary}
\title{Load All Files In Dictionary}
\usage{
load_with_data_dictionary(
  ddFile,
  dbf,
  datadir,
  protocol = NULL,
  dset,
  reppref = F,
  ow = F,
  db = F
)
}
\description{
Load All Files In Dictionary
}
