% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aurum_database.R
\name{make_aurum_database}
\alias{make_aurum_database}
\title{make a full database}
\usage{
make_aurum_database(dbName, dbPath, patient_path, lookup_path, db = F, ow = F)
}
\description{
This is a wrapper for the individual table load functions that loop through each extract table type
and imports the records into one table for each type. It is not meant for clever appending or updating
individual tables it is a either create or overwrite it. If called with overwrite (ow) it will delete
whole database and regenerate.
}
