% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aurum_denomdb.R
\name{make_aurum_denomdb}
\alias{make_aurum_denomdb}
\title{make a demographic linkage database for control construction}
\usage{
make_aurum_denomdb(dbn, denomdir, linkdir, db = F)
}
\description{
Load denominator and linkage tables to enable identification of linkage eligible patients and
construction of controls
}
