% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gold_denomdb.R
\name{make_gold_denomdb}
\alias{make_gold_denomdb}
\title{make a demographic linkage database for control construction}
\usage{
make_gold_denomdb(dbn, denomdir, linkdir, db = F)
}
\description{
make a demographic linkage database for control construction
}
