% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sankey_period_data.R
\name{make_sankey_period_data}
\alias{make_sankey_period_data}
\title{Generate Sankey Network Data}
\usage{
make_sankey_period_data(
  dbf,
  perioddata,
  periods,
  patfilter,
  db = F,
  freq_levels = c("none", "low", "mod", "high"),
  freq_cuts = c(-Inf, 0, 10, 20, Inf),
  pdtab = "temp_perioddata",
  ptab = "temp_periods"
)
}
\description{
This is necessary because of the limitations on SQL Statements to duckdb
}
