% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.R
\name{plot_group}
\alias{plot_group}
\title{Plot a patients time line from timeline data}
\usage{
plot_group(
  data,
  group,
  x_min,
  x_max,
  shapevals = c(Primary = 21, Secondary = 24),
  dbug = F
)
}
\description{
Plot a patients time line from timeline data
}
