% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_execute.R
\name{sql_execute}
\alias{sql_execute}
\title{execute and sql file in a duckdb}
\usage{
sql_execute(dbf, sql_make, sql_query = NULL, sql_drop = NULL, db = F)
}
\description{
execute and sql file in a duckdb
}
