% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_period.R
\name{summarise_period}
\alias{summarise_period}
\title{summarise and event table by periods}
\usage{
summarise_period(
  dbf,
  tabstub = NULL,
  patienttab,
  eventtab,
  idate,
  edate,
  patfilter,
  perstart = NULL,
  perend = NULL,
  perdiv = "months",
  prior = T,
  db = F
)
}
\description{
summarise and event table by periods
}
