% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sql.R
\name{tidy_sql}
\alias{tidy_sql}
\title{tidy up the sql}
\usage{
tidy_sql(
  sql,
  sql_keywords = c("SELECT", "FROM", "WHERE", "AND", "OR", "GROUP", "ORDER", "HAVING",
    "INNER", "LEFT", "RIGHT", "INSERT", "VALUES", "UPDATE", "SET", "DELETE", "CREATE",
    "DROP", "OUTER", "FULL")
)
}
\description{
tidy up the sql
}
