% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank_xml2rdb.R
\name{drugbank_xml2rdb}
\alias{drugbank_xml2rdb}
\title{Parse DrugBank XML to DuckDB (long format, batch inserts)}
\usage{
drugbank_xml2rdb(
  xml_file,
  duckdb_file,
  batchsize = 5000,
  overwrite = FALSE,
  log_every = 500,
  est_total_drugs = 2e+05
)
}
\arguments{
\item{xml_file}{Path to full_database.xml}

\item{duckdb_file}{Path to output .duckdb file}

\item{batchsize}{Number of drugs to batch insert (default 1000)}

\item{overwrite}{Overwrite existing DuckDB file (default FALSE)}
}
\description{
Parse DrugBank XML to DuckDB (long format, batch inserts)
}
