% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_aurum_lookups.R
\name{load_aurum_lookups}
\alias{load_aurum_lookups}
\title{load the aurum lookup tables}
\usage{
load_aurum_lookups(pddir, dbf, ow = F, db = F, silent = T)
}
\arguments{
\item{bdir}{the name of the batch subdirectory}

\item{odir}{the name of the output directory}

\item{olist}{the list of observation codes}

\item{bpp}{BiocParallal Multicore Parameters

Pass a table of covariate codes and generate covariates table}
}
\description{
get the drugissue records and convert some fields to useful field types
}
