% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aurum_database.R
\name{make_aurum_database}
\alias{make_aurum_database}
\title{make a full database}
\usage{
make_aurum_database(dbName, dbPath, patient_path, lookup_path, db = F, ow = F)
}
\description{
make a full database
}
