% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct_make_db.R
\name{sct_make_db}
\alias{sct_make_db}
\title{make a snomed database}
\usage{
sct_make_db(
  dbPath,
  txtPath,
  dbName = NULL,
  db = F,
  ow = T,
  preterm = NULL,
  posterm = NULL,
  prerefs = NULL,
  posrefs = NULL,
  incterm = c("concept", "description", "relationship", "textdefinition"),
  increfs = c("language", "association", "extendedmap", "simplemap", "extendedmap_1",
    "simple")
)
}
\description{
Trying to use some of my existing convenience functions
}
