% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct_make_db.R
\name{sct_make_db}
\alias{sct_make_db}
\title{make a snomed database}
\usage{
sct_make_db(
  dbPath,
  txtPath,
  dbName = NULL,
  db = F,
  ow = T,
  preterm = "sct2_",
  posterm = "_MONOSnap.*",
  prerefs = ".*efset_",
  posrefs = "MONOSnap.*",
  incterm = NULL,
  increfs = NULL
)
}
\description{
Trying to use some of my existing convenience functions
}
