% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sql.R
\name{tidy_sql}
\alias{tidy_sql}
\title{tidy up the sql}
\usage{
tidy_sql(
  sql,
  prefix = "tmp",
  sql_keywords = c("SELECT", "FROM", "WHERE", "AND", "OR", "GROUP", "ORDER", "HAVING",
    "INSERT", "VALUES", "UPDATE", "SET", "DELETE", "CREATE", "DROP", "JOIN", "WITH",
    "INNER", "LEFT", "RIGHT", "OUTER", "FULL", "CROSS", "/\\\\*", "\\\\*/", "--")
)
}
\description{
tidy up the sql
}
